Split plotter UI into companion module
