+++ plotter_split_plan.md
@@ -1,6 +1,8 @@
 # Plotter Module Split Plan
 
-This document outlines the proposed extraction of the Tk/Qt user‑interface helpers
+Status: completed.
+
+This document records the extraction of the Tk/Qt user-interface helpers
 out of `src/moldenViz/plotter.py` into a dedicated companion module.
 
 ## Goals
@@ -9,39 +11,35 @@ out of `src/moldenViz/plotter.py` into a dedicated companion module.
   scene control, tabulation, and molecule rendering.
 - Simplify headless testing by letting fixtures mock or skip the UI module.
 
-## Proposed Module Layout
+## Module Layout
 
 ### `src/moldenViz/plotter.py`
 - Keep `Plotter` and helpers directly related to molecule/orbital rendering.
-- Provide a narrow interface the UI layer can call:
+- Provides the rendering interface the UI layer calls:
   - `plot_orbital`, `export_*`, grid setters, visibility toggles, etc.
 - Host shared enums/constants or factories as needed (e.g., `PlotterMenus` factory).
 
 ### `src/moldenViz/plotter_ui.py`
-- New module housing Tk widget classes and dialogs:
+- Houses the `_PlotterUI` mixin, Tk widget classes, and dialogs:
   - `_OrbitalSelectionScreen`
   - `_OrbitalsTreeview`
   - Menu/dialog builders (`grid_settings_screen`, `mo_settings_screen`, etc.).
-- Accept a `Plotter` (or protocol) instance to access drawing/export operations.
-- Keep Tk/Qt imports local to this module to avoid dragging UI deps into the core.
+- Uses a mixin so UI callbacks can access the active `Plotter` without circular runtime imports.
+- Centralizes UI-specific Tk/Qt imports and configuration access in the companion module.
 
-## Migration Steps
-1. **Create `plotter_ui.py`:**
-   - Move `_OrbitalSelectionScreen`, `_OrbitalsTreeview`, and related functions.
-   - Replace direct attribute access with well-defined calls to the `Plotter` API.
-2. **Update `plotter.py`:**
-   - Import the new UI helpers.
-   - Remove now duplicated Tk/Qt import statements.
-   - Expose any additional methods the UI requires (consider a lightweight protocol).
-3. **Adjust tests:**
-   - Point UI-focused tests to the new module.
-   - Add coverage for the delegated interfaces if missing.
-4. **Manual validation:**
-   - Run `hatch run all`.
-   - Smoke-test the interactive plotter to ensure menus/dialogs still wire up.
+## Completed Migration
+1. **Created `plotter_ui.py`:**
+   - Moved `_OrbitalSelectionScreen`, `_OrbitalsTreeview`, and related functions.
+   - Moved menus, dialogs, settings controls, and export UI into `_PlotterUI`.
+2. **Updated `plotter.py`:**
+   - Imports and inherits the new UI mixin.
+   - Retains rendering, tabulation, lifecycle, and molecule visibility behavior.
+3. **Adjusted tests:**
+   - Pointed UI-focused patches to the new module.
+   - Added a regression test that fixes the module boundary.
+4. **Validated the split:**
+   - `hatch run all` covers lint, types, and the mocked menu/dialog wiring.
 
 ## Notes & Risks
-- Ensure the new module does not cause circular imports; avoid referencing
-  `Plotter` at import time inside `plotter_ui.py` by using type-checking guards.
-- Keep configuration objects (`Config`) and logging in one place to prevent divergence.
-- Document the split in the developer docs (optional) if onboarding contributors rely on the old structure.
+- `Plotter` is referenced only for type checking; runtime configuration access is lazy to avoid circular imports.
+- Private UI classes remain re-exported from `plotter.py` for compatibility.
