+++ pyproject.toml
@@ -38,7 +38,8 @@ dev = [
     "pytest>=8.4",
     "pytest-benchmark>=4.0",
     "pytest-cov>=5.0",
-    "ruff>=0.14",
+    # Ruff 0.15 changes lint defaults and breaks the existing lint baseline.
+    "ruff>=0.14,<0.15",
 ]
 docs = ["pydata-sphinx-theme>=0.16", "sphinx>=8.0"]
